Monkey_Locker_Cervesa/OAuth2/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
httpx
pydantic